# src/NovaLab/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NovaLab API",
    "description": "An ASP.NET Core Web API for managing your streams",
    "version": "v1"
  },
  "paths": {
...
    "/api/twitch/managed-rewards": {
      "get": {
        "tags": [
          "TwitchManagedReward"
        ],
        "operationId": "GetManagedRewards",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include-invalid",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardDtoApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TwitchManagedReward"
        ],
        "operationId": "PostManagedReward",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostManagedRewardDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostManagedRewardDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostManagedRewardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardDtoApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/twitch/managed-rewards/clear": {
      "post": {
        "tags": [
          "TwitchManagedReward"
        ],
        "operationId": "PostNewLastCleared",
        "parameters": [
          {
            "name": "managedRewardId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/twitch/managed-rewards-redemptions": {
      "get": {
        "tags": [
          "TwitchManagedRewardRedemption"
        ],
        "operationId": "GetRedemptions",
        "parameters": [
          {
            "name": "user-id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reward-id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardRedemptionDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardRedemptionDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwitchManagedRewardRedemptionDtoApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TwitchManagedRewardRedemption"
        ],
        "operationId": "PostRedemption",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTwitchManagedRewardRedemptionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostTwitchManagedRewardRedemptionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostTwitchManagedRewardRedemptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateCustomRewardsRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "integer",
            "format": "int32"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "backgroundColor": {
            "type": "string",
            "nullable": true
          },
          "isUserInputRequired": {
            "type": "boolean"
          },
          "isMaxPerStreamEnabled": {
            "type": "boolean"
          },
          "maxPerStream": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isMaxPerUserPerStreamEnabled": {
            "type": "boolean"
          },
          "maxPerUserPerStream": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isGlobalCooldownEnabled": {
            "type": "boolean"
          },
          "globalCooldownSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "shouldRedemptionsSkipRequestQueue": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PostManagedRewardDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "twitchApiRequest": {
            "$ref": "#/components/schemas/CreateCustomRewardsRequest"
          },
          "outputTemplatePerReward": {
            "type": "string",
            "nullable": true
          },
          "outputTemplatePerRedemption": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostTwitchManagedRewardRedemptionDto": {
        "type": "object",
        "properties": {
          "twitchRewardId": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TwitchManagedRewardDto": {
        "type": "object",
        "properties": {
          "managedRewardId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "twitchBroadcasterId": {
            "type": "string",
            "nullable": true
          },
          "twitchRewardId": {
            "type": "string",
            "nullable": true
          },
          "outputTemplatePerRedemption": {
            "type": "string",
            "nullable": true
          },
          "outputTemplatePerReward": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TwitchManagedRewardDtoApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TwitchManagedRewardDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TwitchManagedRewardRedemptionDto": {
        "type": "object",
        "properties": {
          "managedRewardRedemptionId": {
            "type": "string",
            "format": "uuid"
          },
          "managedRewardId": {
            "type": "string",
            "format": "uuid"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TwitchManagedRewardRedemptionDtoApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TwitchManagedRewardRedemptionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}